+++ packages/materials/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.113](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.112...@vtj/materials@0.8.113) (2024-08-23)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.112](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.111...@vtj/materials@0.8.112) (2024-08-22)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.111](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.110...@vtj/materials@0.8.111) (2024-08-22)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.8.110](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.109...@vtj/materials@0.8.110) (2024-08-22)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+
+## [0.8.109](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.108...@vtj/materials@0.8.109) (2024-08-20)
+
+
+### Bug Fixes
+
+* 🐛 field 默认值 ([474824b](https://gitee.com/newgateway/vtj/commits/474824bf3f77f10e89415d87ef377898f06005a9))
+
+
+
+
+
+
 ## [0.8.108](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.8.107...@vtj/materials@0.8.108) (2024-08-18)
 
 **Note:** Version bump only for package @vtj/materials